Add Tooltip Type attribute
The selected type enables different tooltip interfaces.

Form field - The original implementation. The tooltip will be inserted in the embedded form field.
Widget - The tooltip will be applied to the embedded widget.
Standalone - A button with the icon will be inserted into the page, with the tooltip applied.

The default type is 'Form field'.